refactor: extract fetch logic into lib for reuse
Move the ECB exchange rate fetch (+ promise cache) from
app/api/euro/[date]/route.ts into lib/euro-exchange-rate.ts, and move
the Yahoo Finance fetch (+ response cache) from
app/api/stock/[symbol]/daily/route.ts into lib/stock-prices.ts as
fetchSymbolDaily(). Both route handlers become thin wrappers.
